requirements.txt
# AWS SDK (includes Polly, Transcribe, Bedrock)
boto3==1.34.34
botocore==1.34.34

# For local testing
pytest==7.4.3
pytest-mock==3.12.0
moto==4.2.9

# Utilities
python-dateutil==2.8.2
requests==2.31.0

# Streamlit frontend
streamlit==1.31.0
requests==2.31.0
streamlit-audiorecorder>=0.0.6
PyPDF2>=3.0.1
openai>=1.0.0
openai-whisper>=20231117
pydub>=0.25.1
speechrecognition>=3.10.0